Bugfixes, new Files class and new timer class
In this update I overhauled Files class. At first, I removed Exists
function from it, becouse it was unnecessary and then generaly improved
handling of Files. Also created GameTimer class, so we are not using
System.Tickcount anymore. Now we are using Stopwatch. And some more
additions:
- Fixed lua51.dll disappeared
- Fixed sending/receiving NPC data
- Fixed loading of accounts to editors
- Did some changes to ErrorHandler class
- Edited MapNPCs window to be actually child window for Editor suite,
and not new form
- And many many more....