Player Inventories and Npc Placement Changes
- Added Player Inventory Editor (similar to Npc Drop Tables)
- Fixed bug with Npc's displaying one tile below their placement in map
editor
- Added "Npc" layer to dropdown box in Editor Settings - Selected Layer.
- Changed way Npc's are added and removed from the map:
To add an NPC to the map, select the Npc Layer and click where you would
like to place the NPC, this will bring up the selection window to pick
your NPC to place on that tile.
To remove an NPC, select the Npc Layer and right click on the tile your
want to remove the Npc from.
- Added option to turn off Npc rendering in Editor Settings - Visible
Layer's