Map Npc Changes
- Removed MapNpcEditor
- Added ListView with all Npc's names and sprites to the Map Editor
- Moved tileset data to a tab control with the new Npc  list.
- Removed Npc's from EditorProperties -> Selected Layer
- Created new EditorProperty oMode
- Clicking the Tiles tab on the map editor will change you to oMode
Tile, the Npc tab will change you to oMode Npc.
- oModes:
In Tile mode you can left click to place a tile and right click to
remove a tile like normally. You can also use the fill and clear buttons
as normal.
In Npc mode you can select a Npc from the Npc List and use left click to
place that Npc where your cursor is in the map editor and right click
over a current Npc to remove it from the map. You may also use the clear
and fill (although not recommended to use the fill button here) for
Npc's as you would for tiles.
- When you are in Npc mode, your cursor over the map will now display
your currently selected Npc on the map instead of your selected tile.

- Fixed an issue were rendering wasn't properly cleared on the map
editor causing old data to still be rendered.